backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
cloudinary==1.36.0
razorpay==1.3.0
supabase==2.3.0
python-dotenv==1.0.0
cryptography==41.0.8
aiofiles==23.2.1
Pillow==10.1.0
jinja2==3.1.2